changed weights to float value

+++ db/migrate/20210324025444_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[6.1]
     create_table :items do |t|
       t.string :name
       t.string :dis
-      t.integer :weight
+      t.float :weight
 
       t.timestamps
     end
